+++ amethyst_audio/src/systems/dj.rs
@@ -1,13 +1,11 @@
-use std::marker::PhantomData;
-
 use amethyst_assets::AssetStorage;
 use amethyst_core::shred::{Resource, Resources};
 use amethyst_core::specs::common::Errors;
 use amethyst_core::specs::prelude::{Read, System, WriteExpect};
-
 use output::init_output;
 use sink::AudioSink;
 use source::{Source, SourceHandle};
+use std::marker::PhantomData;
 
 /// Calls a closure if the `AudioSink` is empty.
 pub struct DjSystem<F, R> {