+++ amethyst_input/src/lib.rs
@@ -21,14 +21,11 @@ pub use self::button::Button;
 pub use self::controller::{ControllerAxis, ControllerButton};
 pub use self::event::InputEvent;
 pub use self::input_handler::InputHandler;
-pub use self::system::InputSystem;
-pub use self::util::{get_key, is_close_requested, is_key_down};
-
 #[cfg(feature = "sdl_controller")]
 pub use self::sdl_events_system::SdlEventsSystem;
-
+pub use self::system::InputSystem;
+pub use self::util::{get_key, is_close_requested, is_key_down};
 use std::iter::Iterator;
-
 use winit::VirtualKeyCode;
 
 mod axis;