+++ amethyst_renderer/src/error.rs
@@ -1,13 +1,12 @@
 //! Renderer error types.
 
+use amethyst_core;
+use gfx;
+use gfx_core;
 use std::error::Error as StdError;
 use std::fmt::Result as FmtResult;
 use std::fmt::{Display, Formatter};
 use std::result::Result as StdResult;
-
-use amethyst_core;
-use gfx;
-use gfx_core;
 
 /// Renderer result type.
 pub type Result<T> = StdResult<T, Error>;