+++ amethyst_renderer/src/lib.rs
@@ -1,13 +1,31 @@
 //! A data parallel rendering engine developed by the [Amethyst][am] project.
+//!
 //! The source code is available for download on [GitHub][gh]. See the
 //! [online book][bk] for a complete guide to using Amethyst.
+//!
+//! # Background
+//!
+//! This crate provides OpenGL graphics rendering functionality through various *rendering passes*.
+//! The rendering passes may handle different domains of rendering, such as "draw game objects" vs
+//! "render text"; or they can handle the same domain with different variations, such as "draw
+//! objects with lighting" vs "draw objects ignoring lighting".
+//!
+//! ## Skinning: Interleaved Versus Separate Passes
+//!
+//! In an application, objects may be composed of multiple renderable entities, such as a main body
+//! and separate limbs. Where the limbs join the the body, it will look more realistic if the
+//! vertex positions are affected by the relative positions to the body and limb.
+//!
+//! This is where, for a `DrawX` pass, you will find a corresponding `DrawXSeparate` pass which
+//! supports vertex skinning and joint transformations to improve the render. An exception to this
+//! is the `DrawSprite` pass, which does not support joint transformations.
 //!
 //! [am]: https://www.amethyst.rs/
 //! [gh]: https://github.com/amethyst/amethyst/tree/develop/src/renderer
 //! [bk]: https://www.amethyst.rs/book/master/
 
 #![deny(missing_docs)]
-#![doc(html_logo_url = "https://tinyurl.com/jtmm43a")]
+#![doc(html_logo_url = "https://www.amethyst.rs/assets/amethyst.svg")]
 
 extern crate amethyst_assets;
 extern crate amethyst_core;
@@ -25,7 +43,7 @@ extern crate gfx_macros;
 extern crate glsl_layout;
 extern crate hetseq;
 extern crate hibitset;
-extern crate imagefmt;
+extern crate image;
 #[macro_use]
 extern crate log;
 extern crate rayon;
@@ -70,15 +88,16 @@ pub use color::Rgba;
 pub use config::DisplayConfig;
 pub use formats::{
     build_mesh_with_combo, create_mesh_asset, create_texture_asset, BmpFormat, ComboMeshCreator,
-    GraphicsPrefab, ImageData, ImageError, JpgFormat, MaterialPrefab, MeshCreator, MeshData,
-    ObjFormat, PngFormat, TextureData, TextureFormat, TextureMetadata, TexturePrefab,
+    GraphicsPrefab, ImageData, JpgFormat, MaterialPrefab, MeshCreator, MeshData, ObjFormat,
+    PngFormat, TextureData, TextureFormat, TextureMetadata, TexturePrefab,
 };
 pub use input::{ElementState, Event, KeyboardInput, MouseButton, VirtualKeyCode, WindowEvent};
 pub use light::{DirectionalLight, Light, LightPrefab, PointLight, SpotLight, SunLight};
 pub use mesh::{vertex_data, Mesh, MeshBuilder, MeshHandle, VertexBuffer};
 pub use mtl::{Material, MaterialDefaults, MaterialTextureSet, TextureOffset};
 pub use pass::{
-    DrawFlat, DrawFlatSeparate, DrawPbm, DrawPbmSeparate, DrawShaded, DrawShadedSeparate,
+    get_camera, set_vertex_args, DrawFlat, DrawFlatSeparate, DrawPbm, DrawPbmSeparate, DrawShaded,
+    DrawShadedSeparate, DrawSprite,
 };
 pub use pipe::{
     ColorBuffer, Data, DepthBuffer, DepthMode, Effect, EffectBuilder, Init, Meta, NewEffect,
@@ -92,7 +111,10 @@ pub use skinning::{
     AnimatedComboMeshCreator, AnimatedVertexBufferCombination, JointIds, JointTransforms,
     JointTransformsPrefab, JointWeights,
 };
-pub use sprite::{Sprite, SpriteRenderData, SpriteSheet, SpriteSheetHandle, WithSpriteRender};
+pub use sprite::{
+    Sprite, SpriteRender, SpriteSheet, SpriteSheetHandle, SpriteSheetSet, TextureCoordinates,
+};
+pub use sprite_visibility::{SpriteVisibility, SpriteVisibilitySortingSystem};
 pub use system::RenderSystem;
 pub use tex::{Texture, TextureBuilder, TextureHandle};
 pub use transparent::{
@@ -128,6 +150,7 @@ mod resources;
 mod shape;
 mod skinning;
 mod sprite;
+mod sprite_visibility;
 mod system;
 mod tex;
 mod transparent;