+++ amethyst_renderer/src/pass/flat/separate.rs
@@ -1,15 +1,14 @@
 //! Simple flat forward drawing pass.
 
+use super::*;
 use amethyst_assets::AssetStorage;
 use amethyst_core::specs::prelude::{Entities, Join, Read, ReadExpect, ReadStorage};
 use amethyst_core::transform::GlobalTransform;
+use cam::{ActiveCamera, Camera};
+use error::Result;
 use gfx::pso::buffer::ElemStride;
 use gfx_core::state::{Blend, ColorMask};
 use glsl_layout::Uniform;
-
-use super::*;
-use cam::{ActiveCamera, Camera};
-use error::Result;
 use mesh::{Mesh, MeshHandle};
 use mtl::{Material, MaterialDefaults};
 use pass::skinning::{create_skinning_effect, setup_skinning_buffers};
@@ -28,6 +27,13 @@ static ATTRIBUTES: [Attributes<'static>; 2] = [
 ];
 
 /// Draw mesh without lighting
+///
+/// See the [crate level documentation](index.html) for information about interleaved and separate
+/// passes.
+///
+/// # Type Parameters
+///
+/// * `V`: `VertexFormat`
 #[derive(Derivative, Clone, Debug, PartialEq)]
 #[derivative(Default(bound = "Self: Pass"))]
 pub struct DrawFlatSeparate {