+++ examples/appendix_a/pong.rs
@@ -2,11 +2,8 @@ use amethyst::assets::Loader;
 use amethyst::core::cgmath::Vector3;
 use amethyst::core::transform::{GlobalTransform, Transform};
 use amethyst::ecs::prelude::World;
-use amethyst::input::{is_close_requested, is_key_down};
 use amethyst::prelude::*;
-use amethyst::renderer::{
-    Camera, Event, Material, MeshHandle, PosTex, Projection, VirtualKeyCode, WindowMessages,
-};
+use amethyst::renderer::{Camera, Material, MeshHandle, PosTex, Projection, WindowMessages};
 use amethyst::ui::{Anchor, TtfFormat, UiText, UiTransform};
 use config::{ArenaConfig, BallConfig, PaddlesConfig};
 use systems::ScoreText;
@@ -14,7 +11,7 @@ use {Ball, Paddle, Side};
 
 pub struct Pong;
 
-impl<'a, 'b> State<GameData<'a, 'b>> for Pong {
+impl<'a, 'b> SimpleState<'a, 'b> for Pong {
     fn on_start(&mut self, data: StateData<GameData>) {
         let StateData { world, .. } = data;
         use audio::initialise_audio;
@@ -26,19 +23,6 @@ impl<'a, 'b> State<GameData<'a, 'b>> for Pong {
         initialise_audio(world);
         initialise_score(world);
         hide_cursor(world);
-    }
-
-    fn handle_event(&mut self, _: StateData<GameData>, event: Event) -> Trans<GameData<'a, 'b>> {
-        if is_close_requested(&event) || is_key_down(&event, VirtualKeyCode::Escape) {
-            Trans::Quit
-        } else {
-            Trans::None
-        }
-    }
-
-    fn update(&mut self, data: StateData<GameData>) -> Trans<GameData<'a, 'b>> {
-        data.data.update(&data.world);
-        Trans::None
     }
 }